Get rid of ConsoloniaApplication Get rid of ConsoloniaBlazorApplication move intiailization it Consolonia.InitializeConsolonia(application) extension method.
This makes it easier to add a "Console" peer project to an existing avalonia project with multiple hosting application.s.

* Foo
* Foo.Desktop
* Foo.Console

Foo.Console application only needs to derive from Foo's applicationa and it will just work with existing application.
