+++ src/Example.Blazor/App.axaml.cs
@@ -1,9 +1,17 @@
-using Consolonia.Blazor;
+using BlazorBindingsAvalonia;
+using Consolonia;
 using Example.Blazor.Components;
 
 namespace Example.Blazor
 {
-    public class App : ConsoloniaBlazorApplication<MainPage>
+    public class App : BlazorBindingsApplication<MainPage>
     {
+        public override void OnFrameworkInitializationCompleted()
+        {
+            base.OnFrameworkInitializationCompleted();
+
+            this.InitializeConsolonia();
+        }
+
     }
 }
