+++ src/Example/App.cs
@@ -1,11 +1,14 @@
+using Avalonia;
+using Avalonia.Controls.ApplicationLifetimes;
 using Consolonia;
+using Consolonia.Core.Helpers;
 using Consolonia.Themes;
 using Example.Views;
 
 namespace Example
 {
     // ReSharper disable once PartialTypeWithSinglePart
-    public partial class App : ConsoloniaApplication<DataGridTestWindow>
+    public partial class App : Application
     {
         public App()
         {
@@ -14,6 +17,17 @@ namespace Example
             //Styles.Add(new TurboVisionDarkTheme());
             //Styles.Add(new FluentTheme());
             Styles.Add(new MaterialTheme());
+        }
+
+        public override void OnFrameworkInitializationCompleted()
+        {
+            if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
+            {
+                desktop.MainWindow = new DataGridTestWindow();
+                base.OnFrameworkInitializationCompleted();
+
+                this.InitializeConsolonia();
+            }
         }
     }
 }
